requirements.txt
absl-py==0.15.0
aiohttp==3.8.0
aiosignal==1.2.0
albumentations==1.1.0
async-timeout==4.0.0
attrs==21.2.0
cachetools==4.2.4
certifi==2020.12.5
charset-normalizer==2.0.7
colorama==0.4.4
cycler==0.11.0
easydict==1.9
einops==0.3.2
entmax==1.0
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
imageio==2.10.1
joblib==1.1.0
kiwisolver
Markdown==3.3.4
matplotlib
multidict==5.2.0
networkx
oauthlib==3.1.1
olefile==0.46
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==21.2
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytorch-lightning==1.5.3
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image
scikit-learn
scipy
setuptools
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
tifffile
timm
torch
torchmetrics
torchvision
tqdm==4.62.3
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.36.2
wincertstore==0.2
x-transformers==0.20.4
yarl==1.7.2
